requirements.txt
flask
pymupdf
qrcode
pillow
gunicorn